+++ packages/react-native-reanimated/Common/cpp/reanimated/NativeModules/ReanimatedModuleProxy.h
@@ -13,7 +13,6 @@
 #include <reanimated/CSS/registries/CSSKeyframesRegistry.h>
 #include <reanimated/CSS/registries/CSSTransitionsRegistry.h>
 #include <reanimated/CSS/registries/StaticPropsRegistry.h>
-#include <reanimated/Compat/ReactNativeVersionCompat.h>
 #include <reanimated/Compat/WorkletsApi.h>
 #include <reanimated/Events/UIEventHandlerRegistry.h>
 #include <reanimated/Fabric/ReanimatedCommitHook.h>
@@ -30,10 +29,8 @@
 #include <reanimated/Tools/PlatformDepMethodsHolder.h>
 #include <reanimated/Tools/SingleInstanceChecker.h>
 
-#if REACT_NATIVE_VERSION_MINOR >= 85
 #include <react/renderer/animationbackend/AnimationBackend.h>
 #include <react/renderer/uimanager/UIManagerAnimationBackend.h>
-#endif
 
 #include <atomic>
 #include <cstdint>
@@ -209,7 +206,6 @@ class ReanimatedModuleProxy : public std::enable_shared_from_this<ReanimatedModu
   void commitUpdates(const std::unordered_map<SurfaceId, PropsMap> &propsMapBySurface);
   void applySynchronousUpdates(const UpdatesBatch &synchronousUpdatesBatch);
 
-#if REACT_NATIVE_VERSION_MINOR >= 85
   std::shared_ptr<UIManagerAnimationBackend> getAnimationBackend();
   AnimationMutations runGrandCallback(AnimationTimestamp timestamp, GrandCallbackSource source);
   void executeOperationsLoop(AnimationTimestamp timestamp);
@@ -218,7 +214,6 @@ class ReanimatedModuleProxy : public std::enable_shared_from_this<ReanimatedModu
   AnimationMutations collectEventUpdates();
   AnimationMutations collectNonLayoutAnimationUpdates();
   AnimationMutations mutationsFromAnimatedPropsBatch(UpdatesBatchAnimatedProps &&animatedPropsBatch);
-#endif
 
   const bool isReducedMotion_;
   std::atomic<bool> shouldFlushRegistry_{false};
@@ -230,9 +225,7 @@ class ReanimatedModuleProxy : public std::enable_shared_from_this<ReanimatedModu
   RequestRenderFunction requestRender_;
   bool isAnimationRunning_{false};
 
-#if REACT_NATIVE_VERSION_MINOR >= 85
   CallbackId animationBackendCallbackId_{0};
-#endif
 
   // Callbacks queued by OperationsLoop via the requestRender_ override when
   // USE_ANIMATION_BACKEND is on. They are drained at the start of each